+++ greenbridge-junit/src/test/java/com/googlecode/greenbridge/annotation/Story2121.java
@@ -28,7 +28,7 @@ public class Story2121 implements StoryRef {
     }
 
     @Override
-    public int version(){
+    public long version(){
         return 1;
     }
